third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/import-assertions/invalid-import-errors-order-expected.txt

This is a testharness.js-based test.
FAIL Test that the errors order for invalid import declarations is specifier, then attribute key, and then type attribute value. assert_equals: expected function "function SyntaxError() { [native code] }" but got function "function TypeError() { [native code] }"
Harness: the test ran to completion.
